1_bot/requirements.txt
aiofiles==24.1.0
aiogram==2.25.2
aiohappyeyeballs==2.4.4
aiohttp==3.8.6
aiosignal==1.3.2
annotated-types==0.7.0
async-timeout==4.0.3
attrs==24.3.0
Babel==2.9.1
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
frozenlist==1.5.0
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
propcache==0.2.1
psycopg2-binary==2.9.10
pydantic==2.10.5
pydantic_core==2.27.2
pytz==2024.2
requests==2.32.3
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
